requirements.txt
torch
torchvision
numpy
opencv-python
pandas
tqdm
python-chess
Pillow
